binds_p3.rs hands its ctor-field tail and record construct to binds_p3_c.rs, a pure text move under the 800-line ceiling
