+++ crates/almide-mir/src/lower/binds_b.rs
@@ -594,6 +594,7 @@ include!("binds_p2_heap.rs");
 include!("binds_p2_b.rs");
 include!("binds_p2_c.rs");
 include!("binds_p3.rs");
+include!("binds_p3_c.rs");
 include!("binds_p3_b.rs");
 include!("list_literal_elements.rs");
 include!("binds_p4.rs");
