+++ crates/almide-mir/src/lower/binds_p3.rs
@@ -476,421 +476,4 @@ impl LowerCtx {
         }
         self.lower_variant_ctor_field_tail(arg)
     }
-
-    /// The remaining ctor-field classes of [`Self::lower_variant_ctor_field`]:
-    /// Option payloads, closures, and the scalar fallback. Split out so neither
-    /// half outgrows a readable branch ladder; the ORDER across the two halves
-    /// is unchanged (this one runs only after every branch above declined).
-    fn lower_variant_ctor_field_tail(&mut self, arg: &IrExpr) -> Option<(ValueId, bool)> {
-        if let Ty::Applied(almide_lang::types::constructor::TypeConstructorId::Option, a) = &arg.ty {
-            if a.len() == 1 && self.option_payload_drop_exact(&a[0]) {
-                // An Option ctor field whose payload the generated `$__drop_<T>` frees
-                // EXACTLY — scalar (flat rc_dec of the 0-or-1-element block), String /
-                // flat variant (`__drop_list_str` per-element sweep of the 0/1 block),
-                // or a rich variant (`__drop_list_<V>`): the Option block IS a
-                // 0-or-1-element list block, so the drop generator routes it via its
-                // List[T] twin (#1064: `Note{tag: Option[String]}`). A ctor expr
-                // builds the fresh block (`try_lower_option_ctor`); a Var is Dup'd via
-                // `lower_owned_heap_field`. Other payloads (Option[List], Option[record])
-                // stay walled — a later brick, never a leak.
-                let obj = self
-                    .try_lower_option_ctor(arg, &arg.ty)
-                    .or_else(|| self.lower_owned_heap_field(arg))?;
-                return Some((obj, true));
-            }
-            if a.len() == 1 && is_heap_ty(&a[0]) {
-                return None; // un-routable Option payload — decline (wall upstream)
-            }
-        }
-        if matches!(&arg.ty, Ty::Fn { .. }) {
-            // A CLOSURE ctor field (`Run(() => …)` / `Thunk((x) => x * x)` — the
-            // variant-stored closure class): a Lambda arg LIFTS to its closure
-            // block, a Var arg Dups the tracked block (both via
-            // `lower_owned_heap_field`'s existing arms); the ctor then owns the
-            // block and the generated `$__drop_<T>`'s Fn arm frees it via
-            // `__drop_closure` (the classifier + generator admit Fn fields in
-            // the same change — construction and drop agree).
-            let obj = self.lower_owned_heap_field(arg)?;
-            return Some((obj, true));
-        }
-        if is_heap_ty(&arg.ty) {
-            return None; // List[String] / Map / other heap ctor field — a later brick
-        }
-        let v = self.lower_scalar_value(arg)?;
-        Some((v, false))
-    }
-
-    /// Extracted verbatim from `try_lower_variant_ctor` (codopsy round-2 complexity
-    /// sweep, phase 3 of 4): builds the ALL-SCALAR ctor block — tag@slot0, fields@1+,
-    /// zero-filled to the type's uniform `slot_count` — as one target-neutral
-    /// `Op::ListLit`, then applies the same drop/aggregate tracking as the prim path.
-    fn emit_all_scalar_variant_ctor_block(
-        &mut self,
-        tag: i64,
-        slot_count: usize,
-        field_vals: &[(ValueId, bool)],
-        needs_rec: bool,
-        type_name: String,
-    ) -> ValueId {
-        let tagv = self.fresh_value();
-        self.ops.push(Op::ConstInt { dst: tagv, value: tag });
-        let mut slot_vals: Vec<ValueId> = Vec::with_capacity(slot_count);
-        slot_vals.push(tagv);
-        for (v, _) in field_vals {
-            slot_vals.push(*v);
-        }
-        while slot_vals.len() < slot_count {
-            let z = self.fresh_value();
-            self.ops.push(Op::ConstInt { dst: z, value: 0 });
-            slot_vals.push(z);
-        }
-        let dst = self.fresh_value();
-        self.ops.push(Op::ListLit { dst, elems: slot_vals });
-        // EXACT tracking mirror of the prim path (`emit_heap_field_variant_ctor_block` —
-        // heap_slots is empty here, so only the needs_rec branch and the aggregate mark apply).
-        if needs_rec {
-            self.value_drops.entry(dst).or_default().named_route = Some(type_name);
-        }
-        self.materialized_aggregates.insert(dst);
-        dst
-    }
-
-    /// Extracted verbatim from `try_lower_variant_ctor` (codopsy round-2 complexity
-    /// sweep, phase 4 of 4): builds the HEAP-FIELD ctor block on the prim path —
-    /// allocate the `slot_count`-wide block, store the tag then every field slot (heap
-    /// fields moved in), then select the drop route (the recursive `$__drop_<T>` vs
-    /// the masked DropListStr).
-    fn emit_heap_field_variant_ctor_block(
-        &mut self,
-        tag: i64,
-        slot_count: usize,
-        field_vals: Vec<(ValueId, bool)>,
-        needs_rec: bool,
-        type_name: String,
-    ) -> ValueId {
-        use crate::{IntOp, PrimKind};
-        // Allocate the `slot_count`-wide block.
-        let len = self.fresh_value();
-        self.ops.push(Op::ConstInt { dst: len, value: slot_count as i64 });
-        let dst = self.fresh_value();
-        self.ops.push(Op::Alloc {
-            dst,
-            repr: crate::Repr::Ptr { layout: crate::PLACEHOLDER_LAYOUT },
-            init: crate::Init::DynList { len },
-        });
-        let h = self.fresh_value();
-        self.ops.push(Op::Prim { kind: PrimKind::Handle, dst: Some(h), args: vec![dst] });
-        // Store the tag into slot 0, then each field into slot `1+i`. A heap field stores its
-        // HANDLE (i64-widened) then is `Consume`d (moved in); a scalar field stores its value.
-        let tagv = self.fresh_value();
-        self.ops.push(Op::ConstInt { dst: tagv, value: tag });
-        let store_addr = |s: &mut Self, slot: usize| {
-            let off = s.fresh_value();
-            s.ops.push(Op::ConstInt { dst: off, value: layout::slot_offset(slot) as i64 });
-            let addr = s.fresh_value();
-            s.ops.push(Op::IntBinOp { dst: addr, op: IntOp::Add, a: h, b: off });
-            addr
-        };
-        let addr0 = store_addr(self, 0);
-        self.ops.push(Op::Prim { kind: PrimKind::Store { width: 8 }, dst: None, args: vec![addr0, tagv] });
-        let mut heap_slots: Vec<usize> = Vec::new();
-        for (i, (v, is_heap)) in field_vals.into_iter().enumerate() {
-            let slot = 1 + i;
-            let addr = store_addr(self, slot);
-            let store_val = if is_heap {
-                let handle = self.fresh_value();
-                self.ops.push(Op::Prim { kind: PrimKind::Handle, dst: Some(handle), args: vec![v] });
-                handle
-            } else {
-                v
-            };
-            self.ops.push(Op::Prim { kind: PrimKind::Store { width: 8 }, dst: None, args: vec![addr, store_val] });
-            if is_heap {
-                self.ops.push(Op::Consume { v });
-                self.live_heap_handles.retain(|x| *x != v);
-                heap_slots.push(slot);
-            }
-        }
-        // Drop selection: a NESTED-variant type uses the recursive `Op::DropVariant` (the
-        // generated `$__drop_<T>` frees every heap field — variant slots recursively, String
-        // slots flat — then the block). A String-only-field type uses the masked DropListStr.
-        if needs_rec {
-            self.value_drops.entry(dst).or_default().named_route = Some(type_name);
-        } else if !heap_slots.is_empty() {
-            self.record_masks.insert(dst, heap_slots);
-        }
-        self.materialized_aggregates.insert(dst);
-        dst
-    }
-
-    /// Construct a record/tuple with one or more HEAP FIELDS (a `String`/`List`/nested
-    /// aggregate field alongside scalar fields) — `R { name: "x", n: i }`. The block is the
-    /// SAME `[rc][len][cap]` + uniform-i64-slot layout as the scalar path, but each HEAP
-    /// field is a fresh OWNED handle MOVED into its slot (cert `m`), and the value is tracked
-    /// in `record_masks` so its drop frees exactly the heap slots then the block (an
-    /// [`Op::DropListStr`] with the per-value mask — cert = the SAME single `d`).
-    ///
-    /// SOUNDNESS (no new op / no certificate change): this is byte-identical to the
-    /// `List[String]` machinery applied to a mixed slot set. A heap field's owned handle is
-    /// `Consume`d into the slot (cert `m` — moved in, like `prim.store_str`), so each heap
-    /// field is `i…m` (alloc/dup then move-in) and the BLOCK is `i…d` (alloc then the
-    /// recursive `DropListStr`), exactly the balanced shape the proven checker already
-    /// accepts for a list of Strings. A scalar field is a pure `Prim::Store` (no ownership).
-    /// The recursive free at drop touches ONLY the heap slots (the mask) — a scalar slot is
-    /// never `rc_dec`'d. Returns `None` (defer) for an unresolvable type, an omitted heap
-    /// field (a defaulted heap slot would be a garbage handle the drop frees — unsound), or
-    /// a field value not lowerable to an owned handle / scalar.
-    /// A record FIELD's Option/Result CTOR the direct field materializer
-    /// declines (`dp: some([some("a"), none])` — the #1064 codec_field_matrix
-    /// Deep shape): ANF the WHOLE ctor into a tracked synth temp via the full
-    /// `lower_bind` machinery — which registers the temp's TYPED recursive
-    /// drop route — then hand the field the temp as a VAR, so the proven Var
-    /// arm Dups a co-owned ref in. This is byte-for-byte the `let v =
-    /// some(...)` + var-field spelling that already lowers leak-free: the
-    /// record's per-field dec releases its co-ref, and the temp's scope-end
-    /// TYPED drop does the recursion at last ref. (An earlier draft rebuilt
-    /// the ctor inline over an ANF'd payload — the wrapper then belonged to
-    /// the record ALONE, whose generated `$__drop_<R>` flat-decs a
-    /// `List[<recursive heap>]`-normalized field, and the #1530 cap harness
-    /// measured the interior leaking at 152 B/iteration. Ownership shape is
-    /// the fix, not a wider drop arm.) A ctor whose bind DEFERS (Opaque)
-    /// declines — the record then WALLS, never wraps an empty block.
-    fn record_field_ctor_fallback(&mut self, expr: &IrExpr) -> Option<ValueId> {
-        if !matches!(
-            expr.kind,
-            IrExprKind::OptionSome { .. }
-                | IrExprKind::OptionNone
-                | IrExprKind::ResultOk { .. }
-                | IrExprKind::ResultErr { .. }
-        ) {
-            return None;
-        }
-        let tmp = self.fresh_synth_var();
-        if self.lower_bind(tmp, &expr.ty, expr).is_err() {
-            return None;
-        }
-        if self
-            .value_of
-            .get(&tmp)
-            .is_some_and(|v| self.deferred_opaque_binds.contains(v))
-        {
-            return None;
-        }
-        let synth = IrExpr {
-            kind: IrExprKind::Var { id: tmp },
-            ty: expr.ty.clone(),
-            span: expr.span.clone(),
-            def_id: None,
-        };
-        self.lower_owned_heap_field(&synth)
-    }
-
-    pub(crate) fn try_lower_record_construct(&mut self, value: &IrExpr) -> Option<ValueId> {
-        let IrExprKind::Record { fields, .. } = &value.kind else {
-            return None;
-        };
-        // A RECORD-CTOR literal (`Data { payload: …, seq: … }` — the NAME is a registered
-        // variant constructor): this is a TAGGED variant value, NOT a plain record — route
-        // to the variant builder (a tag-less field block here would misread every match).
-        if let IrExprKind::Record { name: Some(n), .. } = &value.kind {
-            if self.variant_layouts.ctor_to_type.contains_key(n.as_str()) {
-                return self.try_lower_variant_ctor(value);
-            }
-        }
-        let Some((names, tys)) = self.aggregate_field_tys(&value.ty) else {
-            crate::trace::trace("ALMIDE_DBG_ELEM", || {
-                format!("[rec-construct] no aggregate layout for ty {:?}", value.ty)
-            });
-            return None;
-        };
-        if tys.is_empty() {
-            return None;
-        }
-        let fields = self.record_fields_with_declared_defaults(&value.ty, &names, fields);
-        let n = tys.len();
-        let heap_slots = self.record_heap_slot_mask(&names, &tys, &fields)?;
-        let slots = self.lower_record_field_slots(&names, &fields)?;
-        let dst = self.emit_record_slot_block(n, slots);
-        self.record_masks.insert(dst, heap_slots);
-        self.materialized_aggregates.insert(dst);
-        if let Some(name) = self.record_drop_type_name(&value.ty) {
-            self.value_drops.entry(dst).or_default().named_route = Some(name);
-        }
-        Some(dst)
-    }
-
-    /// Extracted verbatim from [`Self::try_lower_record_construct`] (codopsy round-3 sweep,
-    /// #852): decides which OMITTED slots a DECLARED default synthesizes as a supplied field.
-    fn record_fields_with_declared_defaults(
-        &self, value_ty: &Ty, names: &[Sym], fields: &[(Sym, IrExpr)],
-    ) -> Vec<(Sym, IrExpr)> {
-        // DEFAULT FILL: an omitted slot with a DECLARED default (`type AllDefault = {
-        // host: String = "localhost", port: Int = 8080 }`; `AllDefault()`) synthesizes
-        // the default as a supplied field — CALL-FREE defaults only (a call default
-        // would inject an uncounted CallFn, breaching the caps mir == ir gate; it
-        // keeps walling via the omitted-heap check below).
-        let mut fields = fields.to_vec();
-        if let Ty::Named(rec_name, _) = value_ty {
-            if let Some(defs) = self
-                .variant_layouts
-                .ctor_field_defaults
-                .get(rec_name.as_str())
-                .cloned()
-            {
-                for nm in names {
-                    if fields.iter().any(|(fname, _)| fname == nm) {
-                        continue;
-                    }
-                    if let Some(d) = defs.get(nm.as_str()) {
-                        if !crate::lower::expr_contains_call(d) {
-                            fields.push((*nm, d.clone()));
-                        }
-                    }
-                }
-            }
-        }
-        fields
-    }
-
-    /// Extracted verbatim from [`Self::try_lower_record_construct`] (codopsy round-3 sweep,
-    /// #852): decides WHICH declared slots hold heap fields — the drop mask — and walls
-    /// (`None`) on an unknown field name, an omitted heap slot, or an all-scalar record.
-    fn record_heap_slot_mask(
-        &self, names: &[Sym], tys: &[Ty], fields: &[(Sym, IrExpr)],
-    ) -> Option<Vec<usize>> {
-        let n = tys.len();
-        // Per-slot heap-ness from the SUPPLIED field's CONCRETE type (`expr.ty`), NOT the
-        // declared field type — a generic field (`first: A` in `Pair[A,B]`) may leave the
-        // DECLARED type an unresolved param that `is_heap_ty` would mis-classify as heap; the
-        // literal's value carries the concrete instantiated type. `None` for an unsupplied
-        // (defaulted) slot — its concrete heap-ness is unknown here.
-        let mut field_heap: Vec<Option<bool>> = vec![None; n];
-        for (name, expr) in fields {
-            let idx = names.iter().position(|n| n == name)?;
-            field_heap[idx] = Some(is_heap_ty(&expr.ty));
-        }
-        // A DEFAULTED (omitted) slot whose DECLARED type is concretely heap (or an unresolved
-        // generic we can't prove scalar) would leave a zero handle the masked drop frees — so
-        // WALL the whole record (never an unsound partial block). A scalar default (a 0 slot)
-        // is fine. (An omitted scalar slot's `field_heap` stays `None` = treated non-heap.)
-        for i in 0..n {
-            if field_heap[i].is_none() && is_heap_ty(&tys[i]) {
-                return None;
-            }
-        }
-        let heap_slots: Vec<usize> =
-            (0..n).filter(|&i| field_heap[i] == Some(true)).collect();
-        if heap_slots.is_empty() {
-            return None; // no heap field — `try_lower_scalar_record_construct` owns it.
-        }
-        Some(heap_slots)
-    }
-
-    /// Extracted verbatim from [`Self::try_lower_record_construct`] (codopsy round-3 sweep,
-    /// #852): lowers every supplied field to its `(declared-index, slot-value, is-heap)`
-    /// triple ahead of the alloc, walling (`None`) on a value it cannot own or copy.
-    fn lower_record_field_slots(
-        &mut self, names: &[Sym], fields: &[(Sym, IrExpr)],
-    ) -> Option<Vec<(usize, ValueId, bool)>> {
-        // Lower each supplied field to (declared-index, slot-value, is-heap). Heap fields
-        // become a fresh OWNED handle (the same kinds `try_lower_str_list_literal` admits);
-        // scalar fields a plain value. All lowered BEFORE the alloc (a field expr that
-        // itself allocates must not interleave with our store sequence).
-        let mut slots: Vec<(usize, ValueId, bool)> = Vec::with_capacity(fields.len());
-        for (name, expr) in fields {
-            let idx = names.iter().position(|n| n == name)?;
-            let is_heap = is_heap_ty(&expr.ty);
-            if is_heap {
-                // A record FIELD additionally takes the FULL Option/Result
-                // ctor set — heap payloads included (`dp: some([some("a"),
-                // none])`, the #1064 codec_field_matrix Deep shape): the
-                // enclosing record's generated `$__drop_<R>` frees every
-                // field by its DECLARED type, so the ctor builder's block is
-                // freed exactly like the `let v = some(...)` + var-field
-                // spelling that already lowers. Deliberately NOT widened in
-                // `lower_owned_heap_field` itself: its other consumers (pair
-                // and list ELEMENT slots) free by flat per-slot rc_dec,
-                // where a heap-payload wrapper would leak its interior.
-                let Some(obj) = self
-                    .lower_owned_heap_field(expr)
-                    .or_else(|| self.record_field_ctor_fallback(expr))
-                else {
-                    crate::trace::trace("ALMIDE_DBG_ELEM", || {
-                        format!(
-                            "[rec-construct] heap field {} ({}) declined",
-                            name.as_str(),
-                            crate::lower::kind_name(&expr.kind)
-                        )
-                    });
-                    return None;
-                };
-                slots.push((idx, obj, true));
-            } else {
-                let Some(v) = self.lower_scalar_value(expr) else {
-                    crate::trace::trace("ALMIDE_DBG_ELEM", || {
-                        format!(
-                            "[rec-construct] scalar field {} ({}) declined",
-                            name.as_str(),
-                            crate::lower::kind_name(&expr.kind)
-                        )
-                    });
-                    return None;
-                };
-                slots.push((idx, v, false));
-            }
-        }
-        Some(slots)
-    }
-
-    /// Extracted verbatim from [`Self::try_lower_record_construct`] (codopsy round-3 sweep,
-    /// #852): emits the `n`-slot block — the alloc plus one `Prim::Store` per slot, each
-    /// heap field's handle stored then `Consume`d (moved in).
-    fn emit_record_slot_block(&mut self, n: usize, slots: Vec<(usize, ValueId, bool)>) -> ValueId {
-        use crate::{IntOp, PrimKind};
-        let len = self.fresh_value();
-        self.ops.push(Op::ConstInt { dst: len, value: n as i64 });
-        let dst = self.fresh_value();
-        self.ops.push(Op::Alloc {
-            dst,
-            repr: crate::Repr::Ptr { layout: crate::PLACEHOLDER_LAYOUT },
-            init: crate::Init::DynList { len },
-        });
-        let h = self.fresh_value();
-        self.ops.push(Op::Prim { kind: PrimKind::Handle, dst: Some(h), args: vec![dst] });
-        for (idx, v, is_heap) in slots {
-            let off = self.fresh_value();
-            self.ops.push(Op::ConstInt { dst: off, value: layout::slot_offset(idx) as i64 });
-            let addr = self.fresh_value();
-            self.ops.push(Op::IntBinOp { dst: addr, op: IntOp::Add, a: h, b: off });
-            // A heap field stores its HANDLE (i64-widened) then is `Consume`d (moved in);
-            // a scalar field stores its value directly.
-            let store_val = if is_heap {
-                let handle = self.fresh_value();
-                self.ops.push(Op::Prim { kind: PrimKind::Handle, dst: Some(handle), args: vec![v] });
-                handle
-            } else {
-                v
-            };
-            self.ops.push(Op::Prim {
-                kind: PrimKind::Store { width: 8 },
-                dst: None,
-                args: vec![addr, store_val],
-            });
-            if is_heap {
-                self.ops.push(Op::Consume { v });
-                self.live_heap_handles.retain(|x| *x != v);
-            }
-        }
-        dst
-    }
-
-    /// Materialize a `List[Record]` LITERAL (`group([rect(…), circle(…)])`, `[el("a"), el("b")]`) — a
-    /// list block whose i64 slots each hold an OWNED Element record handle (lowered via
-    /// `lower_owned_heap_field`, MOVED in). Tracked so its scope-end drop routes to the generated
-    /// `$__drop_list_<R>` (each element freed recursively via `$__drop_<R>`). GATE: the element type
-    /// must be a record needing the recursive drop (`record_drop_type_name` Some), so `$__drop_list_<R>`
-    /// exists; otherwise `None` (the caller keeps the scalar / wall path). Empty lists handled elsewhere.
-    pub(crate) fn try_lower_record_list_literal(&mut self, value: &IrExpr) -> Option<ValueId> {
-        self.try_lower_record_list_literal_as(value, None)
-    }
 }
